Move batch transformations under the transform umbrella
